feat: improve BlockNote paste handling; add markdown detection and parsing; fix race condition with editor initialization
